add Solver class
Solver is used to perform the solution while Graph is used to hold the state of the world
Demands are given to the Solver and no longer part of the Graph
Solve.it! is now a "simple" solve function which fast fails if no solution is found. Use the Solver and Graph classes directly to react to no solution errors